requirements.txt
requests
beautifulsoup4
lxml
pandas
pymongo
python-dotenv==1.0.0
python-box==6.0.2
ensure==1.0.2
joblib
types-PyYAML
pyYAML
aiohttp